src/requirements.txt
kafka-python
pandas
openpyxl
